divisions and play in games can now be created correctly in divisions of 3 and 5

+++ Cribbly/Cribbly/Controllers/GamesController.cs
@@ -47,8 +47,6 @@ namespace Cribbly.Controllers
             var standings = _context.Standings.ToList();
             var divisions = _context.Divisions.ToList();
 
-
-            //WIP: RECONFIGURING CODES TO MAKE GAME NUMBERS AS WELL
             //Combinations so everyone in a 4 team division plays each other once
             int[][] gameCodes = new int[][] 
             {
@@ -73,13 +71,20 @@ namespace Cribbly.Controllers
 
             foreach (var div in divisions)
             {
+                ArgumentOutOfRangeException IsOutOfRange = null;
                 //Get all teams in the division
                 var teams = standings.Where(m => m.Division == div.DivName).ToList();
 
-                try
+                foreach (int[] code in gameCodes)
                 {
-                    foreach (int[] code in gameCodes)
+                    try
                     {
+                        if (teams.Count == 5)
+                        {
+                            IsOutOfRange = new ArgumentOutOfRangeException();
+                            break;
+                        }
+
                         //Create new game
                         var newGame = new PlayInGame()
                         {
@@ -94,16 +99,21 @@ namespace Cribbly.Controllers
                         //Add to DB
                         _context.PlayInGames.Add(newGame);
                     }
+                    catch(ArgumentOutOfRangeException e)
+                    {
+                        IsOutOfRange = e;
+                        break;
+                    }
                 }
-                //Check to ensure that the exception was thrown due to a division with an odd number of teams
-                catch (IndexOutOfRangeException e)
+                    
+                if (IsOutOfRange != null)
                 {
                     //Handle exception based on how many teams are left over
                     var gamesToRemove = _context.PlayInGames.Where(m => m.Division == div.DivName);
 
                     switch (teams.Count)
                     {
-                        case 1:
+                        case 5:
                             //Remove created games, and reconfigure one division for 5 teams
                             _context.PlayInGames.RemoveRange(gamesToRemove);
 
@@ -116,7 +126,7 @@ namespace Cribbly.Controllers
                                     Team1Name = teams[code[0]].TeamName,
                                     Team2Id = teams[code[1]].id,
                                     Team2Name = teams[code[1]].TeamName,
-                                    Division = divisions[0].DivName,
+                                    Division = div.DivName,
                                     GameNumber = code[2],
                                     WinningTeamId = 0
                                 };
@@ -135,71 +145,10 @@ namespace Cribbly.Controllers
                                 Division = div.DivName,
                                 WinningTeamId = 0
                             };
+                            _context.PlayInGames.Add(a);
+
                             break;
-                        case 2:
-                            //Reconfigure two divisions for 5 team divisions
-                            var gamesToRemove2 = _context.PlayInGames.Where(m => m.Division == divisions[0].DivName || m.Division == divisions[1].DivName);
-                            _context.PlayInGames.RemoveRange(gamesToRemove2);
 
-                            foreach (int[] code in div5gameCodes)
-                            {
-                                //Create new game
-                                var newGame = new PlayInGame()
-                                {
-                                    Team1Id = teams[code[0]].id,
-                                    Team1Name = teams[code[0]].TeamName,
-                                    Team2Id = teams[code[1]].id,
-                                    Team2Name = teams[code[1]].TeamName,
-                                    Division = divisions[0].DivName,
-                                    GameNumber = code[2],
-                                    WinningTeamId = 0
-                                };
-                                //Add to DB
-                                _context.PlayInGames.Add(newGame);
-                            }
-
-                            _3WayGame a2 = new _3WayGame()
-                            {
-                                Team1Id = teams[2].id,
-                                Team1Name = teams[2].TeamName,
-                                Team2Id = teams[3].id,
-                                Team2Name = teams[3].TeamName,
-                                Team3Id = teams[4].id,
-                                Team3Name = teams[4].TeamName,
-                                Division = div.DivName,
-                                WinningTeamId = 0
-                            };
-                            var teams2 = standings.Where(m => m.Division == divisions[1].DivName).ToList();
-
-                            foreach (int[] code in div5gameCodes)
-                            {
-                                //Create new game
-                                var newGame = new PlayInGame()
-                                {
-                                    Team1Id = teams2[code[0]].id,
-                                    Team1Name = teams2[code[0]].TeamName,
-                                    Team2Id = teams2[code[1]].id,
-                                    Team2Name = teams2[code[1]].TeamName,
-                                    Division = divisions[1].DivName,
-                                    GameNumber = code[2],
-                                    WinningTeamId = 0
-                                };
-                                //Add to DB
-                                _context.PlayInGames.Add(newGame);
-                            }
-
-                            _3WayGame a3 = new _3WayGame()
-                            {
-                                Team1Id = teams2[2].id,
-                                Team1Name = teams2[2].TeamName,
-                                Team2Id = teams2[3].id,
-                                Team2Name = teams2[3].TeamName,
-                                Team3Id = teams2[4].id,
-                                Team3Name = teams2[4].TeamName,
-                                Division = div.DivName,
-                                WinningTeamId = 0
-                            };
-                            break;
                         case 3:
                             //Configure games for a 3 team division
                             _context.PlayInGames.RemoveRange(gamesToRemove);
@@ -232,11 +181,14 @@ namespace Cribbly.Controllers
                                 Division = div.DivName,
                                 WinningTeamId = 0
                             };
+                            _context.PlayInGames.Add(a4);
+
                             break;
                         default:
                             //Something went wrong making the divisions
                             break;
                     }
+
                 }
             }
             _context.SaveChanges();
